Catch possible connection issues with the MQ
In systems with limited resources, there is a well known problem that celery/rabbitmq cannot handle too many requests. Therefore it's possible that the connection to the MQ fails with errors such as ```ConnectionError: Error 104 while writing to socket. Connection reset by peer.```. In that case, the entire request fails, which is very sad. This commit protects the request by silently failing to register the data to upload.